modified folder structure and file names
