--- AISC2013_Deployment/.gitkeep
@@ -1,2 +1,0 @@
-# T3 AISC2013 - Deployment of AI Solutions
-This folder contains important files.
