--- AISC2013_Deployment/Group_Assignment_1/config.json
@@ -1,4 +1,0 @@
-{
-    "use_tuned_nb": true,
-    "use_tuned_svm": true
-}
