--- AISC2013_Deployment/Group_Assignment_1/test.py
@@ -1,8 +1,0 @@
-import pandas as pd
-
-df = pd.read_csv(r'C:\Users\kiran\Downloads\Yelp JSON\review.csv')
-
-print(df)
-
-print(df.loc[0])
-print(df.loc[0,'text'])
