--- Computer Vision/README.md
@@ -1,1 +1,0 @@
-# AISC-T3-Grp1
