+++ Deployment/Assignment 1/config.json
@@ -1,0 +1,4 @@
+{
+    "use_tuned_nb": true,
+    "use_tuned_svm": true
+}
