+++ Deployment/Assignment 1/test.py
@@ -1,0 +1,8 @@
+import pandas as pd
+
+df = pd.read_csv(r'C:\Users\kiran\Downloads\Yelp JSON\review.csv')
+
+print(df)
+
+print(df.loc[0])
+print(df.loc[0,'text'])
